First commit; created a small proxy to fetch images for a user using the Facbook SDK.
Currently, one must manually generate a user access token via the Graph Explorer API and
store it in a 'secrets.json' file keyed as 'access_token'. The same must be done with an
App-id ('app_id') and App-secret ('app_secret'). This file name is then passed to the proxy
which handles parsing it and using the information to create an API connection.

I added a 'retrieveAccessToken()' method for future implementation, as I would like to circumvent
having to manually create an access token if at all possible.